requirements.txt
asgiref==3.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
boto3==1.20.31
botocore==1.21.65
cement==2.8.2
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
decorator==5.1.0
Django==4.0
django-environ==0.8.1
future==0.16.0
idna==3.3
ipdb==0.13.9
ipython==7.30.1
jedi==0.18.1
jmespath==0.10.0
matplotlib-inline==0.1.3
numpy==1.21.5
parso==0.8.3
pathspec==0.5.9
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.24
Pygments==2.10.0
pypiwin32==223
python-dateutil==2.8.2
pywin32==303
PyYAML==5.4.1
requests==2.26.0
s3transfer==0.5.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
termcolor==1.1.0
timm==0.4.12
toml==0.10.2
torch==1.10.1+cu102
torchvision==0.11.2+cu102
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
wcwidth==0.2.5
